:recycle: (demo) Move checkBluetoothPermissions in dedicated service file
